+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2022-01-24 - Edu
+### Changed
+- **forks/close_quotes.c** Dinamyc CTRL+D error single double quotes open
+
 ## 2022-01-23 - Edu
 ### Added
 - **libft/ft_free.c** frees a pointer & set as NULL